02_Docker/read_api/requirements.txt
requests>=2.31.0
minio>=7.2.5
python-dotenv>=1.0.1